Adding Hotkeys to quickly run each autochecklist
"CONTROL + SHIFT + A" will run the AutoCheckList ULP for the BOARD.

"CONTROL + SHIFT + S" will run the AutoCheckList ULP for the SCHEMATIC.

Note, these ULPs are still in progress, but they are currently checking 9 of the line items on the Engineering Design Rule Checklist. With some work, we could automate many more of the line items, and hopefully dramatically reduce the amount of line items that must be checked manually.