Prioritized PROD_ID to indicate 'BOM' status first.
Found that my external display indicator was identified as non-hardware because it has no hardware on the board, despite having a PROD_ID. Changed so that if there is a PROD_ID then it is prioritized. Not sure how this will affect other situations, but let's give it a shot.